requirement.txt
Flask
mysql-connector-python

pip install -r requirement.txt